refactor: consolidate language construction in queries package
also split queries/queries.go into smaller, focused files
