+++ queries/queries.go
@@ -18,34 +18,14 @@ package queries
 
 import (
 	"embed"
-	"errors"
 	"fmt"
-	"iter"
-	"path"
-	"slices"
 	"strings"
-	"sync"
-	"time"
 
-	"github.com/pterm/pterm"
-
-	tsBlade "github.com/EmranMR/tree-sitter-blade/bindings/go"
 	ts "github.com/tree-sitter/go-tree-sitter"
-	tsCss "github.com/tree-sitter/tree-sitter-css/bindings/go"
-	tsET "github.com/tree-sitter/tree-sitter-embedded-template/bindings/go"
-	tsHtml "github.com/tree-sitter/tree-sitter-html/bindings/go"
-	tsJsdoc "github.com/tree-sitter/tree-sitter-jsdoc/bindings/go"
-	tsPHP "github.com/tree-sitter/tree-sitter-php/bindings/go"
-	tsTypescript "github.com/tree-sitter/tree-sitter-typescript/bindings/go"
-
-	tsHandlebars "bennypowers.dev/tree-sitter-handlebars/bindings/go"
-	tsJinja "bennypowers.dev/tree-sitter-jinja-dialects/bindings/go"
 )
 
 //go:embed */*.scm
 var queries embed.FS
-
-var ErrNoQueryManager = errors.New("QueryManager is nil")
 
 type NoCaptureError struct {
 	Capture string
@@ -55,95 +35,6 @@ type NoCaptureError struct {
 func (e *NoCaptureError) Error() string {
 	return fmt.Sprintf("No nodes for capture %s in query %s", e.Capture, e.Query)
 }
-
-// ---- Languages struct holds pre-initialized language grammars ----
-var languages = struct {
-	typescript       *ts.Language
-	jsdoc            *ts.Language
-	css              *ts.Language
-	html             *ts.Language
-	tsx              *ts.Language
-	blade            *ts.Language
-	jinja            *ts.Language
-	handlebars       *ts.Language
-	embeddedTemplate *ts.Language
-	php              *ts.Language
-}{
-	ts.NewLanguage(tsTypescript.LanguageTypescript()),
-	ts.NewLanguage(tsJsdoc.Language()),
-	ts.NewLanguage(tsCss.Language()),
-	ts.NewLanguage(tsHtml.Language()),
-	ts.NewLanguage(tsTypescript.LanguageTSX()),
-	ts.NewLanguage(tsBlade.Language()),
-	ts.NewLanguage(tsJinja.Language()),
-	ts.NewLanguage(tsHandlebars.Language()),
-	ts.NewLanguage(tsET.Language()),
-	ts.NewLanguage(tsPHP.LanguagePHP()),
-}
-
-// ---- Parser Pooling Section ----
-
-func newParserPool(lang *ts.Language) *sync.Pool {
-	return &sync.Pool{
-		New: func() any {
-			parser := ts.NewParser()
-			if err := parser.SetLanguage(lang); err != nil {
-				panic(fmt.Sprintf("failed to set language: %v", err))
-			}
-			return parser
-		},
-	}
-}
-
-func getParser(pool *sync.Pool) *ts.Parser {
-	return pool.Get().(*ts.Parser)
-}
-
-func putParser(pool *sync.Pool, parser *ts.Parser) {
-	parser.Reset()
-	pool.Put(parser)
-}
-
-var (
-	htmlParserPool       = newParserPool(languages.html)
-	cssParserPool        = newParserPool(languages.css)
-	jsdocParserPool      = newParserPool(languages.jsdoc)
-	typescriptParserPool = newParserPool(languages.typescript)
-	tsxParserPool        = newParserPool(languages.tsx)
-	bladeParserPool      = newParserPool(languages.blade)
-	jinjaParserPool      = newParserPool(languages.jinja)
-	hbsParserPool        = newParserPool(languages.handlebars)
-	etParserPool         = newParserPool(languages.embeddedTemplate)
-	phpParserPool        = newParserPool(languages.php)
-)
-
-// Note: Removed cursor pooling due to stateful nature of QueryCursor objects.
-// Cursors maintain internal state that can affect subsequent queries,
-// leading to unpredictable behavior. Always create fresh cursors.
-
-func GetHTMLParser() *ts.Parser              { return getParser(htmlParserPool) }
-func PutHTMLParser(parser *ts.Parser)         { putParser(htmlParserPool, parser) }
-func GetCSSParser() *ts.Parser                { return getParser(cssParserPool) }
-func PutCSSParser(parser *ts.Parser)          { putParser(cssParserPool, parser) }
-func GetJSDocParser() *ts.Parser              { return getParser(jsdocParserPool) }
-func PutJSDocParser(parser *ts.Parser)        { putParser(jsdocParserPool, parser) }
-
-func RetrieveTypeScriptParser() *ts.Parser     { return getParser(typescriptParserPool) }
-func PutTypeScriptParser(parser *ts.Parser)    { putParser(typescriptParserPool, parser) }
-func GetTSXParser() *ts.Parser                 { return getParser(tsxParserPool) }
-func PutTSXParser(parser *ts.Parser)           { putParser(tsxParserPool, parser) }
-func GetBladeParser() *ts.Parser               { return getParser(bladeParserPool) }
-func PutBladeParser(parser *ts.Parser)         { putParser(bladeParserPool, parser) }
-func GetJinjaParser() *ts.Parser               { return getParser(jinjaParserPool) }
-func PutJinjaParser(parser *ts.Parser)         { putParser(jinjaParserPool, parser) }
-func GetHandlebarsParser() *ts.Parser          { return getParser(hbsParserPool) }
-func PutHandlebarsParser(parser *ts.Parser)    { putParser(hbsParserPool, parser) }
-func GetEmbeddedTemplateParser() *ts.Parser    { return getParser(etParserPool) }
-func PutEmbeddedTemplateParser(parser *ts.Parser) { putParser(etParserPool, parser) }
-func GetPHPParser() *ts.Parser                 { return getParser(phpParserPool) }
-func PutPHPParser(parser *ts.Parser)           { putParser(phpParserPool, parser) }
-
-// ---- End Parser Pooling Section ----
 
 // QuerySelector defines which queries to load for performance
 type QuerySelector struct {
@@ -187,177 +78,6 @@ func GenerateQueries() QuerySelector {
 	}
 }
 
-type QueryManagerI interface {
-	Close()
-	getQuery(name string) (*ts.Query, error)
-}
-
-type QueryManager struct {
-	typescript map[string]*ts.Query
-	jsdoc      map[string]*ts.Query
-	css        map[string]*ts.Query
-	html       map[string]*ts.Query
-	tsx        map[string]*ts.Query
-	blade      map[string]*ts.Query
-}
-
-func NewQueryManager(selector QuerySelector) (*QueryManager, error) {
-	start := time.Now()
-	qm := &QueryManager{
-		typescript: make(map[string]*ts.Query),
-		jsdoc:      make(map[string]*ts.Query),
-		css:        make(map[string]*ts.Query),
-		html:       make(map[string]*ts.Query),
-		tsx:        make(map[string]*ts.Query),
-		blade:      make(map[string]*ts.Query),
-	}
-
-	// Load only the requested queries
-	for _, queryName := range selector.HTML {
-		if err := qm.loadQuery("html", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load HTML query %s: %w", queryName, err)
-		}
-	}
-
-	for _, queryName := range selector.TypeScript {
-		if err := qm.loadQuery("typescript", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load TypeScript query %s: %w", queryName, err)
-		}
-	}
-
-	for _, queryName := range selector.CSS {
-		if err := qm.loadQuery("css", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load CSS query %s: %w", queryName, err)
-		}
-	}
-
-	for _, queryName := range selector.JSDoc {
-		if err := qm.loadQuery("jsdoc", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load JSDoc query %s: %w", queryName, err)
-		}
-	}
-
-	for _, queryName := range selector.TSX {
-		if err := qm.loadQuery("tsx", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load TSX query %s: %w", queryName, err)
-		}
-	}
-
-	for _, queryName := range selector.Blade {
-		if err := qm.loadQuery("blade", queryName); err != nil {
-			qm.Close()
-			return nil, fmt.Errorf("failed to load Blade query %s: %w", queryName, err)
-		}
-	}
-
-	pterm.Debug.Println("Constructing selected queries took", time.Since(start))
-	return qm, nil
-}
-
-func (qm *QueryManager) loadQuery(language, queryName string) error {
-	// Blade reuses html/*.scm files compiled against the Blade grammar
-	sourceDir := language
-	if language == "blade" {
-		sourceDir = "html"
-	}
-	// Use path.Join (not filepath.Join) - embed.FS requires POSIX / separators
-	queryPath := path.Join(sourceDir, queryName+".scm")
-	data, err := queries.ReadFile(queryPath)
-	if err != nil {
-		return fmt.Errorf("failed to read query file %s: %w", queryPath, err)
-	}
-
-	queryText := string(data)
-	var tsLang *ts.Language
-	switch language {
-	case "typescript":
-		tsLang = languages.typescript
-	case "jsdoc":
-		tsLang = languages.jsdoc
-	case "css":
-		tsLang = languages.css
-	case "html":
-		tsLang = languages.html
-	case "tsx":
-		tsLang = languages.tsx
-	case "blade":
-		tsLang = languages.blade
-	default:
-		return fmt.Errorf("unknown language %s", language)
-	}
-
-	query, qerr := ts.NewQuery(tsLang, queryText)
-	if qerr != nil {
-		return fmt.Errorf("failed to parse query %s: %w", queryName, qerr)
-	}
-
-	switch language {
-	case "typescript":
-		qm.typescript[queryName] = query
-	case "jsdoc":
-		qm.jsdoc[queryName] = query
-	case "css":
-		qm.css[queryName] = query
-	case "html":
-		qm.html[queryName] = query
-	case "tsx":
-		qm.tsx[queryName] = query
-	case "blade":
-		qm.blade[queryName] = query
-	}
-
-	return nil
-}
-
-func (qm *QueryManager) Close() {
-	for _, query := range qm.typescript {
-		query.Close()
-	}
-	for _, query := range qm.jsdoc {
-		query.Close()
-	}
-	for _, query := range qm.css {
-		query.Close()
-	}
-	for _, query := range qm.html {
-		query.Close()
-	}
-	for _, query := range qm.tsx {
-		query.Close()
-	}
-	for _, query := range qm.blade {
-		query.Close()
-	}
-}
-
-func (qm *QueryManager) getQuery(queryName string, language string) (*ts.Query, error) {
-	var q *ts.Query
-	var ok bool
-	switch language {
-	case "typescript":
-		q, ok = qm.typescript[queryName]
-	case "jsdoc":
-		q, ok = qm.jsdoc[queryName]
-	case "css":
-		q, ok = qm.css[queryName]
-	case "html":
-		q, ok = qm.html[queryName]
-	case "tsx":
-		q, ok = qm.tsx[queryName]
-	case "blade":
-		q, ok = qm.blade[queryName]
-	}
-	if !ok {
-		return nil, fmt.Errorf("unknown query %s", queryName)
-	}
-	return q, nil
-}
-
 type ParentNodeCaptures struct {
 	NodeId   uintptr
 	Captures CaptureMap
@@ -371,143 +91,6 @@ type CaptureInfo struct {
 }
 
 type CaptureMap = map[string][]CaptureInfo
-
-type QueryMatcher struct {
-	query  *ts.Query
-	cursor *ts.QueryCursor
-}
-
-func (qm QueryMatcher) Close() {
-	// NOTE: we don't close queries here, only at the end of execution in QueryManager.Close
-	// Close the cursor since we're not pooling
-	qm.cursor.Close()
-}
-
-func (qm QueryMatcher) GetCaptureNameByIndex(index uint32) string {
-	return qm.query.CaptureNames()[index]
-}
-
-func (qm QueryMatcher) CaptureCount() int {
-	return len(qm.query.CaptureNames())
-}
-
-func (qm QueryMatcher) GetCaptureIndexForName(name string) (uint, bool) {
-	return qm.query.CaptureIndexForName(name)
-}
-
-func (qm QueryMatcher) SetByteRange(start uint, end uint) {
-	qm.cursor.SetByteRange(start, end)
-}
-
-func NewQueryMatcher(
-	manager *QueryManager,
-	language string,
-	queryName string,
-) (*QueryMatcher, error) {
-	if manager == nil {
-		return nil, ErrNoQueryManager
-	}
-	query, err := manager.getQuery(queryName, language)
-	if err != nil {
-		return nil, err
-	}
-	cursor := ts.NewQueryCursor()
-	qm := QueryMatcher{query, cursor}
-	return &qm, nil
-}
-
-func (q QueryMatcher) AllQueryMatches(node *ts.Node, text []byte) iter.Seq[*ts.QueryMatch] {
-	qm := q.cursor.Matches(q.query, node, text)
-	return func(yield func(qm *ts.QueryMatch) bool) {
-		for {
-			m := qm.Next()
-			if m == nil {
-				break
-			}
-			if !yield(m) {
-				return
-			}
-		}
-	}
-}
-
-// ParentCaptures returns an iterator over unique parent node captures as identified by the given parent capture name.
-// For each unique parent node (e.g., a field or method), it aggregates all captures from all query matches sharing
-// that parent node into a single CaptureMap. This allows you to collect all related captures (such as attributes,
-// decorators, etc.) for each parent node in the source AST.
-//
-// Example usage:
-//
-//	for captures := range matcher.ParentCaptures(root, code, "field") {
-//	  // captures represents the captured nodes for a single field
-//	  addFieldToDeclaration(captures, declaration)
-//	}
-func (q *QueryMatcher) ParentCaptures(root *ts.Node, code []byte, parentCaptureName string) iter.Seq[CaptureMap] {
-	names := q.query.CaptureNames()
-
-	type pgroup struct {
-		capMap    CaptureMap
-		startByte uint
-	}
-
-	parentGroups := make(map[int]pgroup)
-
-	for match := range q.AllQueryMatches(root, code) {
-		var parentNode *ts.Node
-		for _, cap := range match.Captures {
-			name := names[cap.Index]
-			if name == parentCaptureName {
-				parentNode = &cap.Node
-				break
-			}
-		}
-		if parentNode == nil {
-			continue
-		}
-		pid := int(parentNode.Id())
-		startByte := parentNode.StartByte()
-		_, ok := parentGroups[pid]
-		if !ok {
-			capmap := make(CaptureMap)
-			parentGroups[pid] = pgroup{capmap, startByte}
-		}
-		for _, cap := range match.Captures {
-			name := names[cap.Index]
-			text := cap.Node.Utf8Text(code)
-			ci := CaptureInfo{
-				NodeId:    int(cap.Node.Id()),
-				Text:      text,
-				StartByte: cap.Node.StartByte(),
-				EndByte:   cap.Node.EndByte(),
-			}
-			if _, hasMap := parentGroups[pid].capMap[name]; !hasMap {
-				parentGroups[pid].capMap[name] = make([]CaptureInfo, 0)
-			}
-			if !slices.ContainsFunc(parentGroups[pid].capMap[name], func(m CaptureInfo) bool {
-				return m.NodeId == ci.NodeId
-			}) {
-				parentGroups[pid].capMap[name] = append(parentGroups[pid].capMap[name], ci)
-			}
-		}
-	}
-
-	sorted := make([]pgroup, 0)
-	for _, group := range parentGroups {
-		sorted = append(sorted, group)
-	}
-
-	slices.SortStableFunc(sorted, func(a pgroup, b pgroup) int {
-		return int(a.startByte) - int(b.startByte)
-	})
-
-	return func(yield func(CaptureMap) bool) {
-		for _, group := range sorted {
-			if !yield(group.capMap) {
-				break
-			}
-		}
-	}
-}
 
 func GetDescendantById(root *ts.Node, id int) *ts.Node {
 	c := root.Walk()
@@ -544,6 +127,18 @@ type Range struct {
 	End   Position
 }
 
+// NodeToRange converts a tree-sitter node to a Range using byte-to-position conversion
+func NodeToRange(node *ts.Node, content []byte) Range {
+	start := byteOffsetToPosition(content, node.StartByte())
+	end := byteOffsetToPosition(content, node.EndByte())
+	return Range{
+		Start: start,
+		End:   end,
+	}
+}
+
+// Helper functions
+
 // byteOffsetToPosition converts a byte offset to line/character position
 func byteOffsetToPosition(content []byte, offset uint) Position {
 	line := uint32(0)
@@ -567,184 +162,6 @@ func byteOffsetToPosition(content []byte, offset uint) Position {
 		Character: char,
 	}
 }
-
-// NodeToRange converts a tree-sitter node to a Range using byte-to-position conversion
-func NodeToRange(node *ts.Node, content []byte) Range {
-	start := byteOffsetToPosition(content, node.StartByte())
-	end := byteOffsetToPosition(content, node.EndByte())
-	return Range{
-		Start: start,
-		End:   end,
-	}
-}
-
-// FindClassDeclarationInSource finds the class declaration position in TypeScript source content
-func FindClassDeclarationInSource(content []byte, className string, queryManager *QueryManager) (*Range, error) {
-	// Get TypeScript parser and parse
-	parser := RetrieveTypeScriptParser()
-	defer PutTypeScriptParser(parser)
-
-	tree := parser.Parse(content, nil)
-	if tree == nil {
-		return nil, nil
-	}
-	defer tree.Close()
-
-	// Get class queries
-	classQueries, err := NewQueryMatcher(queryManager, "typescript", "classes")
-	if err != nil {
-		return nil, err
-	}
-	defer classQueries.Close()
-
-	// Find class declaration using ParentCaptures pattern
-	for captureMap := range classQueries.ParentCaptures(tree.RootNode(), content, "class") {
-		if classNames, ok := captureMap["class.name"]; ok {
-			for _, classNameCapture := range classNames {
-				if classNameCapture.Text == className {
-					node := GetDescendantById(tree.RootNode(), classNameCapture.NodeId)
-					if node != nil {
-						r := NodeToRange(node, content)
-						return &r, nil
-					}
-				}
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-// FindTagNameDefinitionInSource finds where the tag name is defined (e.g., @customElement decorator)
-func FindTagNameDefinitionInSource(content []byte, tagName string, queryManager *QueryManager) (*Range, error) {
-	// Get TypeScript parser and parse
-	parser := RetrieveTypeScriptParser()
-	defer PutTypeScriptParser(parser)
-
-	tree := parser.Parse(content, nil)
-	if tree == nil {
-		return nil, nil
-	}
-	defer tree.Close()
-
-	// Get class queries
-	classQueries, err := NewQueryMatcher(queryManager, "typescript", "classes")
-	if err != nil {
-		return nil, err
-	}
-	defer classQueries.Close()
-
-	// Find tag name in decorator using ParentCaptures pattern
-	for captureMap := range classQueries.ParentCaptures(tree.RootNode(), content, "class") {
-		if tagNames, ok := captureMap["tag-name"]; ok {
-			for _, tagNameCapture := range tagNames {
-				if tagNameCapture.Text == tagName {
-					node := GetDescendantById(tree.RootNode(), tagNameCapture.NodeId)
-					if node != nil {
-						r := NodeToRange(node, content)
-						return &r, nil
-					}
-				}
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-// FindAttributeDeclarationInSource finds the attribute/property declaration in source content
-func FindAttributeDeclarationInSource(content []byte, attributeName string, queryManager *QueryManager) (*Range, error) {
-	// Get TypeScript parser and parse
-	parser := RetrieveTypeScriptParser()
-	defer PutTypeScriptParser(parser)
-
-	tree := parser.Parse(content, nil)
-	if tree == nil {
-		return nil, nil
-	}
-	defer tree.Close()
-
-	// Get member queries
-	memberQueries, err := NewQueryMatcher(queryManager, "typescript", "classMemberDeclaration")
-	if err != nil {
-		return nil, err
-	}
-	defer memberQueries.Close()
-
-	// Find attribute declaration using ParentCaptures pattern
-	for captureMap := range memberQueries.ParentCaptures(tree.RootNode(), content, "field") {
-		var memberName string
-		var attributeNameInDecorator string
-		var memberNodeId int
-
-		// Collect information from this field
-		if memberNames, ok := captureMap["member.name"]; ok && len(memberNames) > 0 {
-			memberName = memberNames[0].Text
-			memberNodeId = memberNames[0].NodeId
-		}
-		if attrNames, ok := captureMap["field.attr.name"]; ok && len(attrNames) > 0 {
-			attributeNameInDecorator = attrNames[0].Text
-		}
-
-		// Check if this member matches our target attribute
-		if matchesAttribute(memberName, attributeNameInDecorator, attributeName) {
-			node := GetDescendantById(tree.RootNode(), memberNodeId)
-			if node != nil {
-				r := NodeToRange(node, content)
-				return &r, nil
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-// FindSlotDefinitionInSource finds the slot definition in template within source content
-func FindSlotDefinitionInSource(content []byte, slotName string, queryManager *QueryManager) (*Range, error) {
-	// Get TypeScript parser and parse
-	parser := RetrieveTypeScriptParser()
-	defer PutTypeScriptParser(parser)
-
-	tree := parser.Parse(content, nil)
-	if tree == nil {
-		return nil, nil
-	}
-	defer tree.Close()
-
-	// Get class queries to find render templates
-	classQueries, err := NewQueryMatcher(queryManager, "typescript", "classes")
-	if err != nil {
-		return nil, err
-	}
-	defer classQueries.Close()
-
-	// Find HTML template strings in render methods
-	for captureMap := range classQueries.ParentCaptures(tree.RootNode(), content, "class") {
-		if templates, ok := captureMap["render.template"]; ok {
-			for _, template := range templates {
-				templateContent := template.Text
-				// Remove template literal backticks
-				if len(templateContent) >= 2 && templateContent[0] == '`' && templateContent[len(templateContent)-1] == '`' {
-					templateContent = templateContent[1 : len(templateContent)-1]
-				}
-
-				slotRange := findSlotInTemplate(templateContent, slotName, queryManager)
-				if slotRange != nil {
-					// Adjust range to account for position within the TypeScript file
-					templateNode := GetDescendantById(tree.RootNode(), template.NodeId)
-					if templateNode != nil {
-						adjustedRange := adjustTemplateRange(slotRange, templateNode, content)
-						return adjustedRange, nil
-					}
-				}
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-// Helper functions
 
 func matchesAttribute(memberName, decoratorAttrName, targetAttr string) bool {
 	// If decorator explicitly sets attribute name, use that
@@ -824,98 +241,4 @@ func adjustTemplateRange(templateRange *Range, templateNode *ts.Node, content []
 			Character: templateStart.Character + templateRange.End.Character,
 		},
 	}
-}
-
-// Thread-safe singleton QueryManager (there can be only one!)
-var (
-	globalQueryManager *QueryManager
-	globalQueryOnce    sync.Once
-	globalQueryError   error
-)
-
-// GetGlobalQueryManager returns the singleton QueryManager instance
-func GetGlobalQueryManager() (*QueryManager, error) {
-	globalQueryOnce.Do(func() {
-		manager, err := NewQueryManager(LSPQueries())
-		if err != nil {
-			globalQueryError = err
-			return
-		}
-		globalQueryManager = manager
-	})
-
-	if globalQueryError != nil {
-		return nil, globalQueryError
-	}
-
-	if globalQueryManager == nil {
-		return nil, fmt.Errorf("failed to initialize global query manager")
-	}
-
-	return globalQueryManager, nil
-}
-
-// GetCachedQueryMatcher returns a query matcher with cached query but fresh cursor
-// This prevents concurrent access to the same cursor which causes segmentation faults
-func GetCachedQueryMatcher(manager *QueryManager, language, queryName string) (*QueryMatcher, error) {
-	if manager == nil {
-		return nil, ErrNoQueryManager
-	}
-
-	// Get the cached query (thread-safe to share)
-	query, err := manager.getQuery(queryName, language)
-	if err != nil {
-		return nil, err
-	}
-
-	// Always create a fresh cursor (NOT thread-safe to share)
-	cursor := ts.NewQueryCursor()
-	matcher := QueryMatcher{query, cursor}
-	return &matcher, nil
-}
-
-// FindDefinedElementTags uses the definedElements.scm query to find
-// custom element tag names defined in TypeScript/JavaScript source code
-// via @customElement decorators or customElements.define calls.
-func FindDefinedElementTags(code []byte, qm *QueryManager) []string {
-	parser := RetrieveTypeScriptParser()
-	defer PutTypeScriptParser(parser)
-
-	tree := parser.Parse(code, nil)
-	if tree == nil {
-		return nil
-	}
-	defer tree.Close()
-
-	root := tree.RootNode()
-	if root == nil {
-		return nil
-	}
-
-	// Returns nil on error — consistent with the nil-return convention for
-	// parse failures above. Callers treat nil as "no definitions found".
-	matcher, err := GetCachedQueryMatcher(qm, "typescript", "definedElements")
-	if err != nil {
-		return nil
-	}
-	defer matcher.Close()
-
-	var tags []string
-	for match := range matcher.AllQueryMatches(root, code) {
-		if match == nil {
-			continue
-		}
-		for _, capture := range match.Captures {
-			if int(capture.Index) >= matcher.CaptureCount() {
-				continue
-			}
-			if matcher.GetCaptureNameByIndex(capture.Index) == "defined.tagName" {
-				tag := capture.Node.Utf8Text(code)
-				if !slices.Contains(tags, tag) {
-					tags = append(tags, tag)
-				}
-			}
-		}
-	}
-	return tags
 }
